+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/skill_gems/SkillGemsAdder.java
@@ -1,5 +1,8 @@
 package com.robertx22.age_of_exile.aoe_data.database.skill_gems;
 
+import com.robertx22.age_of_exile.aoe_data.database.skill_gems.support.DexSupportGems;
+import com.robertx22.age_of_exile.aoe_data.database.skill_gems.support.IntSupportGems;
+import com.robertx22.age_of_exile.aoe_data.database.skill_gems.support.StrSupportGems;
 import com.robertx22.age_of_exile.database.registry.Database;
 import com.robertx22.age_of_exile.database.registry.ISlashRegistryInit;
 
@@ -8,7 +11,9 @@ public class SkillGemsAdder implements ISlashRegistryInit {
     @Override
     public void registerAll() {
 
-        new SupportSkillGemsAdder().registerAll();
+        new IntSupportGems().registerAll();
+        new StrSupportGems().registerAll();
+        new DexSupportGems().registerAll();
 
         Database.Spells()
             .getSerializable()