+++ src/main/java/com/robertx22/age_of_exile/aoe_data/datapacks/lang_file/CreateLangFile.java
@@ -114,6 +114,8 @@ public class CreateLangFile {
 
         list.addAll(Database.MobAffixes()
             .getSerializable());
+        list.addAll(Database.SkillGemRarities()
+            .getSerializable());
         list.addAll(Database.SkillGems()
             .getSerializable());
         list.addAll(Database.Perks()