+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/bonus_dmg_to_status_affected/BonusDmgToStatusAffected.java
@@ -14,7 +14,7 @@ public class BonusDmgToStatusAffected extends Stat {
 
     public static BonusDmgToStatusAffected FROST = new BonusDmgToStatusAffected(NegativeEffects.FROSTBURN, "Chilled", "chill");
     public static BonusDmgToStatusAffected BURN = new BonusDmgToStatusAffected(NegativeEffects.BURN, "Burning", "burn");
-    public static BonusDmgToStatusAffected POISON = new BonusDmgToStatusAffected(NegativeEffects.THORNS, "Poisoned", "poison");
+    public static BonusDmgToStatusAffected POISON = new BonusDmgToStatusAffected(NegativeEffects.POISON, "Poisoned", "poison");
     public static BonusDmgToStatusAffected STATIC = new BonusDmgToStatusAffected(NegativeEffects.SHOCK, "Charged", "static");
 
     String affectedName;