+++ src/main/java/com/robertx22/age_of_exile/uncommon/enumclasses/Elements.java
@@ -14,7 +14,8 @@ public enum Elements implements IColor {
     Water(2, new RGB(0, 128, 255), true, "Ice", Formatting.AQUA, "water", "\u2749"),
     Thunder(3, new RGB(204, 0, 255), true, "Lightning", Formatting.YELLOW, "thunder", "\u272A"),
     Nature(4, new RGB(0, 204, 0), true, "Poison", Formatting.DARK_GREEN, "nature", "\u273F"),
-    Elemental(5, new RGB(0, 0, 0), false, "Elemental", Formatting.GOLD, "elemental", "\u269C");
+    Elemental(5, new RGB(0, 0, 0), false, "Elemental", Formatting.GOLD, "elemental", "\u269C"),
+    All(6, new RGB(0, 204, 0), false, "All", Formatting.LIGHT_PURPLE, "all", "\u273F");
 
     public boolean isSingleElement = true;
     private RGB color;
@@ -67,6 +68,10 @@ public enum Elements implements IColor {
         }
 
         if (this == other) {
+            return true;
+        }
+
+        if (other == All || this == All) {
             return true;
         }