+++ src/main/java/com/robertx22/age_of_exile/vanilla_mc/items/SkillGemItem.java
@@ -66,7 +66,9 @@ public class SkillGemItem extends Item implements IAutoLocName, IAutoModel {
             SkillGem gem = data.getSkillGem();
 
             tooltip.clear();
-            tooltip.add(gem.locName());
+            tooltip.add(gem.locName()
+                .formatted(data.getRarity()
+                    .textFormatting()));
 
             tooltip.addAll(data.GetTooltipString(new TooltipInfo(ClientOnly.getPlayer())));